file modulated
